==> NSI_BE/2-catalog.ipynb <==
import matplotlib as mpl
import matplotlib.pyplot as plt
import pandas as pd
import json
import requests

from sentinelhub import (
    CRS,
    BBox,
    DataCollection,
    MimeType,
    MosaickingOrder,
    SentinelHubRequest,
    SentinelHubStatistical,
    SHConfig,
)

config = SHConfig("cdse")

config.sh_token_url = "https://identity.dataspace.copernicus.eu/auth/realms/CDSE/protocol/openid-connect/token"
config.sh_base_url = "https://sh.dataspace.copernicus.eu"

for collection in DataCollection:
    print(collection)

from oauthlib.oauth2 import BackendApplicationClient
from requests_oauthlib import OAuth2Session

# Create a session
client = BackendApplicationClient(client_id=config.sh_client_id)
oauth = OAuth2Session(client=client)

# Get token for the session
token = oauth.fetch_token(token_url='https://identity.dataspace.copernicus.eu/auth/realms/CDSE/protocol/openid-connect/token',
                          client_secret=config.sh_client_secret, 
                          include_client_id=True)

# All requests using this session will have an access token automatically added
resp = oauth.get("https://sh.dataspace.copernicus.eu/configuration/v1/wms/instances")
print(resp.content)

# Get collections

# Get collections
url = "https://sh.dataspace.copernicus.eu/api/v1/catalog/1.0.0/collections"
response = oauth.get(url)

if response.status_code == 200:
    catalog=response.json()

for collection in catalog['collections']:
    print(f"{collection['id']:<20}: {collection['description']}")

import json

print(json.dumps(catalog, indent=4))

# Inspect a collection

collection='sentinel-2-l2a'
url = f"https://sh.dataspace.copernicus.eu/api/v1/catalog/1.0.0/collections/{collection}/"
response = oauth.get(url)

if response.status_code == 200:
    catalog=response.json()

for band in catalog["summaries"]["eo:bands"]:
    print(band)

print(json.dumps(catalog, indent=4))

# Search within a collection

# Search format for sentinelhub package
bbox = BBox((3924000,3081000,3937000,3094000), crs=CRS(3035)).transform(3035)
time_interval =("2024-12-08","2024-12-08")
resolution = ((3937000-3924000)/100, (3094000-3081000)/100)

# Search format for catalog api
# Example
data = {
    "bbox": [13, 45, 14, 46],
    "datetime": "2019-12-10T00:00:00Z/2019-12-10T23:59:59Z",
    "collections": ["sentinel-1-grd"],
    "limit": 5,
}

bbox = BBox((3924000,3081000,3937000,3094000), crs=CRS(3035)).transform(CRS.WGS84)
bbox=[bbox.min_x,bbox.min_y,bbox.max_x,bbox.max_y]

date_start="2024-01-01"
date_end="2024-12-31"

collection='sentinel-2-l2a'

data = {
    "bbox": bbox,
    "datetime": f"{date_start}T00:00:00Z/{date_end}T23:59:59Z",
    "collections": [collection],
    "limit": 5,
}

url = "https://sh.dataspace.copernicus.eu/api/v1/catalog/1.0.0/search"
response = oauth.post(url, json=data)

print(response.status_code)

print(json.dumps(response.json(), indent=4))

# Search a BYOC

data = {
    "bbox": [13, 45, 14, 46],
    "datetime": "2019-12-10T00:00:00Z/2019-12-10T23:59:59Z",
    "collections": ["byoc-<byoc-collection-id>"],
    "limit": 5,
}

url = "https://sh.dataspace.copernicus.eu/api/v1/catalog/1.0.0/search"
response = oauth.post(url, json=data)

# "36caf873-32ca-44be-ba3e-e6c1b759211c" communiqué lors des présentations d'Eurostat
byoc_id="36caf873-32ca-44be-ba3e-e6c1b759211c"
date_start="2000-01-01"
date_end="2024-12-31"
data = {
    "bbox": bbox,
    "datetime": f"{date_start}T00:00:00Z/{date_end}T23:59:59Z",
    "collections": [f"byoc-{byoc_id}"],
    "limit": 5,
}

url = "https://sh.dataspace.copernicus.eu/api/v1/catalog/1.0.0/search"
response = oauth.post(url, json=data)

if response.status_code == 200:
    catalog=response.json()
else:
    print(f'ERROR: {response.status_code}')

import json

print(json.dumps(catalog, indent=4))
