==> NSI_BE/5-parquet-concat.ipynb <==
import pandas as pd
import os

# Concatenate all parquet files

Purpose: distribute a single large file to colleagues, instead of multiple smaller files.

# Concaténer le long des lignes (axe=0)
concat = pd.concat([pd.read_parquet(f'parquet/{file}') for file in os.listdir('parquet')], axis=0)

concat.to_parquet('data/sentinel_2018.parquet')
