==> NSI_BE/5-parquet2bboxes.ipynb <==
import pandas as pd
import os

# Find bboxes of all parquet files

Open all parquet files and find bboxes.
Structure of output dataset: file-(min_x,min_y)-(max_x,max_y)

with open('data/summary.psv', 'w') as file:
    file.write('file|min_x|min_y|max_x|max_y\n')

    for file in os.listdir('parquet'):
        df=pd.read_parquet(f'parquet/{file}')
        for row in duckdb.sql('''
            select min(X_rounded) as min_x, 
                   min(Y_rounded) as min_y, 
                   max(X_rounded) as max_x, 
                   max(Y_rounded) as max_y
            from df
            ''').to_df().itertuples():
            file.write(f'{name}|{row.min_x}|{row.min_y}|{row.max_x}|{row.max_y}\n')

summary=pd.read_csv('data_2024/summary.psv',sep='|')

summary

# Fucntion to identify file that cover given coordinates

import duckdb
def getFile(x,y):
    return duckdb.sql(f'''
    select file
    from summary
    where min_x<={x} and {x}<=max_x
     and  min_y<={y} and {y}<=max_y 
     ''').to_df()
    
print(getFile(3930000,3154400))
