==> NSI_BE/6-extracts for QGis.ipynb <==
import pandas as pd
import duckdb
import os
import shutil

codes=pd.read_csv('corine/TD_LAND_COVER.csv',sep='|')
codes.to_parquet('corine/TD_LAND_COVER.parquet',index=False)

file='013356689bd3389f3f29430412edb6a9_2018'

identifier,year=file.split('_')

%%time

extract=duckdb.sql(f'''
select *,
       'POLYGON ((' || X        || ' ' || Y         || ', ' ||
                      (X + 100) || ' ' || Y         || ', ' ||
                      (X + 100) || ' ' || (Y + 100) || ', ' ||
                       X        || ' ' || (Y + 100) || ', ' ||
                       X        || ' ' || Y         || '))' AS wkt_polygon
from (
      select sentinel.X_rounded as X,
             sentinel.Y_rounded as Y,
             sentinel.*,
             codes.*
      from 'parquet_{year}_quantiles/{file}.parquet' as sentinel,
           'corine/clc_benelux_grid_BL.parquet'      as corine,
           'corine/TD_LAND_COVER.parquet'            as codes
      where corine.X=sentinel.X_rounded
      and   corine.Y=sentinel.Y_rounded
      and   corine.VALUE=codes.CD_COVER          
      )
''').to_df()
extract.to_csv(f'visualisations/{file}_statistics_with_clc.csv',sep='|',index=False)

extract.shape

duckdb.sql(f'''
            select '{file}'           as file,
                   min(X_rounded)     as min_x, 
                   min(Y_rounded)     as min_y, 
                   max(X_rounded)+100 as max_x, 
                   max(Y_rounded)+100 as max_y
            from extract
''').to_df()

shutil.copyfile(f'data_{year}/{identifier}/response.tiff',f'visualisations/{file}.tiff')
